Got basic license wizard working.
Removed some unneeded files (tabinterface.*)
updated all the test class (removed the qDebug() statements)
reorganized the license class and added code to the wizard to
make the basic wizard function, accept valid input and save it
to the file. WIP, but the basics are in place.
Added image to the wizard.
Cleaned up all the warnings.