Overhauled the settingsui. Updated settings by name
Rewrote the code for the settings. Now all you have to do to create
a setting is to add it to the list in the settings.cpp file and add
the ui element with the same name to the settings.ui file. the code
in the settingsui.cpp file will automatically connect the two items.