Major cleanup of functionality for stitch mode.
Single click adds a cell to the chart.
if there is a selection clicking on the chart removes selection only.
only use/create a rubberband if the user has dragged the right distance.

Also only draw a rowLine if we start with a cell, and only delete it if
there is a line to start with.