Fix bugs in group/ungroup behavior.
This patch fixes a couple of bugs. First when you group a set of stitches
you shouldn't be able to select the children when using a rubberband, etc.
Second When ungrouping a set of stitches you need to make sure you can
select them, and that they are alread selected.